+++ package.json
@@ -74,7 +74,11 @@
         "pretty-quick": "^3.1.3"
     },
     "lint-staged": {
-        "*.{js,ts,jsx,tsx}": ["prettier ./src --write", "eslint ./src", "git add ."]
+        "*.{js,ts,jsx,tsx}": [
+            "prettier ./src --write",
+            "eslint ./src",
+            "git add ."
+        ]
     },
     "husky": {
         "hooks": {
