+++ src/App.test.tsx
@@ -1,9 +1,14 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import App from './App';
+import GlobalStateProvider from './global';
 
 test('renders learn react link', () => {
-    render(<App />);
-    const linkElement = screen.getByText(/learn react/i);
-    expect(linkElement).toBeInTheDocument();
+    render(
+        <GlobalStateProvider>
+            <App />
+        </GlobalStateProvider>
+    );
+    const RouterComponent = screen.getByTestId('Router');
+    expect(RouterComponent).toBeInTheDocument();
 });
